+++ actors/test/multisig_delete_self_test.go
@@ -64,8 +64,7 @@ func TestMultisigDeleteSelf2Of3RemovedIsProposer(t *testing.T) {
 	vm.ApplyOk(t, v, addrs[1], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams)
 
 	// txnid not found when third approval gets processed indicating that the transaction has gone through successfully
-	result, err := v.ApplyMessage(addrs[2], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
-	require.NoError(t, err)
+	result := vm.RequireApplyMessage(t, v, addrs[2], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
 	assert.Equal(t, exitcode.ErrNotFound, result.Code)
 
 }
@@ -116,8 +115,7 @@ func TestMultisigDeleteSelf2Of3RemovedIsApprover(t *testing.T) {
 	vm.ApplyOk(t, v, addrs[0], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams)
 
 	// txnid not found when third approval gets processed indicating that the transaction has gone through successfully
-	result, err := v.ApplyMessage(addrs[2], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
-	require.NoError(t, err)
+	result := vm.RequireApplyMessage(t, v, addrs[2], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
 	assert.Equal(t, exitcode.ErrNotFound, result.Code)
 
 }
@@ -166,8 +164,7 @@ func TestMultisigDeleteSelf2Of2(t *testing.T) {
 	vm.ApplyOk(t, v, addrs[1], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams)
 
 	// txnid not found when another approval gets processed indicating that the transaction has gone through successfully
-	result, err := v.ApplyMessage(addrs[1], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
-	require.NoError(t, err)
+	result := vm.RequireApplyMessage(t, v, addrs[1], multisigAddr, big.Zero(), builtin.MethodsMultisig.Approve, &approveRemoveSignerParams, t.Name())
 	assert.Equal(t, exitcode.ErrNotFound, result.Code)
 }
