+++ support/vm/testing.go
@@ -304,7 +304,7 @@ func AdvanceByDeadline(t *testing.T, v *VM, minerIDAddr address.Address, predica
 		v, err = v.WithEpoch(dlInfo.Last())
 		require.NoError(t, err)
 
-		result, err := v.ApplyMessage(builtin.SystemActorAddr, builtin.CronActorAddr, big.Zero(), builtin.MethodsCron.EpochTick, nil, t.Name())
+		result := RequireApplyMessage(t, v, builtin.SystemActorAddr, builtin.CronActorAddr, big.Zero(), builtin.MethodsCron.EpochTick, nil, t.Name())
 		require.NoError(t, err)
 		require.Equal(t, exitcode.Ok, result.Code)
 
@@ -481,10 +481,15 @@ func GetDealState(t *testing.T, vm *VM, dealID abi.DealID) (*market.DealState, b
 //
 
 func ApplyOk(t *testing.T, v *VM, from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}) cbor.Marshaler {
-	result, err := v.ApplyMessage(from, to, value, method, params, t.Name())
-	require.NoError(t, err)
+	result := RequireApplyMessage(t, v, from, to, value, method, params, t.Name())
 	require.Equal(t, exitcode.Ok, result.Code, "unexpected exit code")
 	return result.Ret
+}
+
+func RequireApplyMessage(t *testing.T, v *VM, from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}, name string) MessageResult {
+	result, err := v.ApplyMessage(from, to, value, method, params, name)
+	require.NoError(t, err)
+	return result
 }
 
 //
