+++ support/vm/vm.go
@@ -3,10 +3,7 @@ package vm
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
 	"fmt"
-	"os"
-	"strings"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-state-types/abi"
@@ -343,6 +340,8 @@ type MessageResult struct {
 	GasCharged int64
 }
 
+// ApplyMessage applies the message to the current state. It returns result of message application and any internal vm errors.
+// If test-vector environment variables are set this method generates tests-vectors as a side effect
 func (vm *VM) ApplyMessage(from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}, info string) (MessageResult, error) {
 	vectorGen := newVectorGen()
 
@@ -350,16 +349,17 @@ func (vm *VM) ApplyMessage(from, to address.Address, value abi.TokenAmount, meth
 		return MessageResult{}, err
 	}
 
-	result, callSeq, fakesAccessed := vm.applyMessageInternal(from, to, value, method, params)
+	result, callSeq, fakesAccessed, err := vm.applyMessageInternal(from, to, value, method, params)
+	if err != nil {
+		return MessageResult{}, err
+	}
 	if err := vectorGen.after(vm, from, to, value, method, params, callSeq, result, fakesAccessed, info); err != nil {
 		return MessageResult{}, err
 	}
 	return result, nil
 }
 
-// ApplyMessage applies the message to the current state. It returns result of message application, callSeqNum
-// used during application and whether a fake syscall was accessed during invocation
-func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}) (MessageResult, uint64, bool) {
+func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}) (MessageResult, uint64, bool, error) {
 	// This method does not actually execute the message itself,
 	// but rather deals with the pre/post processing of a message.
 	// (see: `invocationContext.invoke()` for the dispatch and execution)
@@ -368,16 +368,16 @@ func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmou
 	// load actor from global state
 	fromID, ok := vm.NormalizeAddress(from)
 	if !ok {
-		return MessageResult{nil, exitcode.SysErrSenderInvalid, gasCharged}, 0, false
+		return MessageResult{nil, exitcode.SysErrSenderInvalid, gasCharged}, 0, false, nil
 	}
 
 	fromActor, found, err := vm.GetActor(fromID)
 	if err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 	if !found {
 		// Execution error; sender does not exist at time of message execution.
-		return MessageResult{nil, exitcode.SysErrSenderInvalid, gasCharged}, 0, false
+		return MessageResult{nil, exitcode.SysErrSenderInvalid, gasCharged}, 0, false, nil
 	}
 
 	// send
@@ -390,7 +390,7 @@ func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmou
 	callSeq := fromActor.CallSeqNum
 	fromActor.CallSeqNum = callSeq + 1
 	if err := vm.setActor(context.Background(), fromID, fromActor); err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 
 	// checkpoint state
@@ -398,16 +398,16 @@ func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmou
 	// - CallSeqNumber increment
 	priorRoot, err := vm.checkpoint()
 	if err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 
 	msg, err := makeChainMessage(from, to, callSeq, value, method, params)
 	if err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 	var msgBuf bytes.Buffer
 	if err := msg.MarshalCBOR(&msgBuf); err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 	bs := msgBuf.Bytes()
 	charge := vm.gasPrices.OnChainMessage(len(bs))
@@ -449,24 +449,24 @@ func (vm *VM) applyMessageInternal(from, to address.Address, value abi.TokenAmou
 	// proceed from a nested call failure.
 	if exitCode != exitcode.Ok {
 		if err := vm.rollback(priorRoot); err != nil {
-			panic(err)
+			return MessageResult{}, 0, false, err
 		}
 	} else {
 		// persist changes from final invocation if call is ok
 		if _, err := vm.checkpoint(); err != nil {
-			panic(err)
+			return MessageResult{}, 0, false, err
 		}
 	}
 
 	// serialize return and charge gas
 	var retBuf bytes.Buffer
 	if err := ret.inner.MarshalCBOR(&retBuf); err != nil {
-		panic(err)
+		return MessageResult{}, 0, false, err
 	}
 	retGasCharge := vm.gasPrices.OnChainReturnValue(len(retBuf.Bytes()))
 	gasCharged = retGasCharge.Total() + ctx.topLevel.gasUsed
 
-	return MessageResult{ret.inner, exitCode, gasCharged}, callSeq, ctx.topLevel.fakeSyscallsAccessed
+	return MessageResult{ret.inner, exitCode, gasCharged}, callSeq, ctx.topLevel.fakeSyscallsAccessed, nil
 }
 
 func (vm *VM) StateRoot() cid.Cid {
@@ -708,120 +708,4 @@ func (msg InternalMessage) Caller() address.Address {
 // Receiver implements runtime.MessageInfo.
 func (msg InternalMessage) Receiver() address.Address {
 	return msg.to
-}
-
-//
-// Test Vector generation utilities
-//
-
-type vectorGen struct {
-	conformanceDir string
-	determinismDir string
-	vector         TestVector
-}
-
-func newVectorGen() *vectorGen {
-	// check environment variables to determine if generation is on
-	conformanceDir := os.Getenv("SPECS_ACTORS_CONFORMANCE")
-	determinismDir := os.Getenv("SPECS_ACTORS_DETERMINISM")
-	return &vectorGen{
-		conformanceDir: conformanceDir,
-		determinismDir: determinismDir,
-	}
-}
-
-func (g *vectorGen) determinism() bool {
-	return g.determinismDir != ""
-}
-
-func (g *vectorGen) conformance() bool {
-	return g.conformanceDir != ""
-}
-
-func (g *vectorGen) before(v *VM, info string) error {
-	if g.determinism() || g.conformance() {
-		// Set test vector pre application conditions
-		startOpts := StartConditions(v, info)
-		for _, opt := range startOpts {
-			if err := opt(&(g.vector)); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func (g *vectorGen) after(v *VM, from, to address.Address, value abi.TokenAmount, method abi.MethodNum, params interface{}, callSeq uint64, result MessageResult, fakesAccessed bool, info string) error {
-	if !g.conformance() && !g.determinism() {
-		return nil
-	}
-	// Set test vector message and post application conditions
-	if err := SetMessage(from, to, callSeq, value, method, params)(&(g.vector)); err != nil {
-		return err
-	}
-	if err := SetEndStateTree(v.StateRoot(), v.store)(&(g.vector)); err != nil {
-		return err
-	}
-	if err := SetReceipt(result)(&(g.vector)); err != nil {
-		return err
-	}
-	b, err := (&(g.vector)).MarshalJSON()
-	if err != nil {
-		return err
-	}
-
-	fromID, _ := v.NormalizeAddress(from)
-	toID, _ := v.NormalizeAddress(to)
-	act, _, _ := v.GetActor(toID)
-	actName := strings.Split(builtin.ActorNameByCode(act.Code), "/")[2]
-
-	h := sha256.Sum256(b)
-	fname := fmt.Sprintf("%x-%s-%s-%s-%d.json", string(h[:]), fromID, toID, actName, method)
-
-	// Write conformance test-vectors
-	if g.conformance() && !fakesAccessed {
-		dir := g.conformanceDir + "/" + info
-		exists, err := dirExists(dir)
-		if err != nil {
-			return err
-		}
-		if !exists {
-			if err := os.MkdirAll(dir, 0755); err != nil {
-				return err
-			}
-		}
-		if err := os.WriteFile(dir+"/"+fname, b, 0755); err != nil {
-			return err
-		}
-	}
-
-	// Write determinism test-vectors
-	if g.determinism() {
-		dir := g.determinismDir + "/" + info
-		exists, err := dirExists(dir)
-		if err != nil {
-			return err
-		}
-		if !exists {
-			if err := os.MkdirAll(dir, 0755); err != nil {
-				return err
-			}
-		}
-		if err := os.WriteFile(dir+"/"+fname, b, 0755); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// dirExists returns whether the given file or directory exists
-func dirExists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err == nil {
-		return true, nil
-	}
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-	return false, err
 }
